Removed "Porter" directory from unit and integration test hierarchy.
